backend/requirements-clean.txt
"""
Copyright (c) 2025 Kevin Zachary
All rights reserved.

This software and associated documentation files (the "Software") are the 
exclusive property of Kevin Zachary. Unauthorized copying, distribution, 
modification, or use of this software is strictly prohibited.

For licensing inquiries, contact: kevin@zachary.com
"""

# Core FastAPI dependencies
fastapi>=0.109.0
uvicorn[standard]==0.22.0
sqlalchemy>=1.4,<1.5
databases>=0.6.0,<0.7.0
asyncpg==0.27.0
pydantic>=2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.31.0
httpx==0.24.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
aiohttp==3.8.5

# Authentication dependencies
authlib>=1.2.0
starlette>=0.27.0

# Lightweight ML dependencies (CPU-only)
scikit-learn==1.3.0
numpy==1.26.4
pandas==2.0.3
joblib
statsmodels
shap>=0.41.0
torch>=2.0.0

# Lightweight transformers (without CUDA dependencies)
transformers==4.31.0
tokenizers==0.13.3

# Other dependencies
itsdangerous==2.1.2
pydantic-settings>=2.0.0
PyYAML>=6.0.1
APScheduler>=3.10.0
openai
grpcio==1.56.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-cloud-core>=2.0.0
google-cloud-compute>=1.0.0
google-cloud-storage>=2.0.0

# Development tools
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
isort==5.12.0
psutil>=5.9.0

# Slack integration
slack-sdk==3.21.3
